Bug fix/learning track (#906)
* fix

* fix:blog component
