--- app/(misson center)/mission-center/layout.tsx
@@ -1,7 +1,0 @@
-import { Metadata } from 'next';
-import Page from './page';
-
-export const metadata: Metadata = {
-  title: 'Mission Center'
-};
-export default Page;
