+++ app/(resource station)/blog/[blogId]/page.tsx
@@ -1,44 +1,19 @@
-'use client';
-import React, { useEffect } from 'react';
-import BlogHeader from './BlogHeader';
-import BlogContent from './BlogContent';
-import BlogFooter from './BlogFooter';
-import { BurialPoint } from '@/helper/burialPoint';
-import { useParams } from 'next/navigation';
-import { useRequest } from 'ahooks';
-import webApi from '@/service';
-import Loading from '@/components/Common/Loading';
-import { BlogDetailType } from '@/service/webApi/resourceStation/type';
+import { FC } from 'react';
+import BlogDetail from '../components/BlogId';
+import { Metadata } from 'next';
 
-interface BlogDetailProp {}
+export const metadata: Metadata = {
+  title: 'HackQuest Blog Detail'
+};
 
-const BlogDetail: React.FC<BlogDetailProp> = () => {
-  const { blogId } = useParams();
-  const { data: blog = {} as BlogDetailType, loading } = useRequest(
-    async () => {
-      const res = await webApi.resourceStationApi.getBlogDetail(
-        blogId as string
-      );
-      return res;
-    }
-  );
-  useEffect(() => {
-    const startTime = new Date().getTime();
-    return () => {
-      const endTime = new Date().getTime();
-      const duration = endTime - startTime;
-      BurialPoint.track('blog-content-page-页面留存时间', {
-        duration
-      });
-    };
-  }, []);
+interface BlogPageProps {}
+
+const BlogPage: FC<BlogPageProps> = (props) => {
   return (
-    <Loading loading={loading} className="font-next-book text-[16px]">
-      <BlogHeader blog={blog} />
-      <BlogContent blog={blog} />
-      <BlogFooter />
-    </Loading>
+    <>
+      <BlogDetail />
+    </>
   );
 };
 
-export default BlogDetail;
+export default BlogPage;
