+++ app/(resource station)/blog/components/BlogContent/index.tsx
@@ -1,8 +1,8 @@
 'use client';
 import React, { useMemo } from 'react';
 import { BlogDetailType } from '@/service/webApi/resourceStation/type';
-import ComponentRender from '../components/ComponentRender';
 import { CustomComponent } from '@/components/Web/Business/Renderer/type';
+import ComponentRender from '../ComponentRender';
 
 interface BlogContentProp {
   blog: BlogDetailType;
