--- app/(resource station)/blog/layout.tsx
@@ -1,7 +1,0 @@
-import { Metadata } from 'next';
-import Page from './page';
-
-export const metadata: Metadata = {
-  title: 'HackQuest Blog'
-};
-export default Page;
