+++ app/mobile/(resource station)/blog/components/BlogFooter/index.tsx
