+++ app/mobile/(resource station)/blog/components/BlogHeader/index.tsx
